# stackhawk-openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "PrintFlow API",
  "description": "Store-scoped print order queue: CSV intake, PSD composition, TIFF/PDF output, CUPS dispatch.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/users": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "List Users",
    "operationId": "list_users_users_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "store_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Store Id"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/UserOut"
         },
         "title": "Response List Users Users Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Create User",
    "operationId": "create_user_users_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "parameters": []
   }
  },
  "/users/{user_id}": {
   "patch": {
    "tags": [
     "users"
    ],
    "summary": "Update User",
    "operationId": "update_user_users__user_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "users"
    ],
    "summary": "Delete User",
    "operationId": "delete_user_users__user_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Role": {
    "type": "string",
    "enum": [
     "ADMIN",
     "OPERATOR"
    ],
    "title": "Role"
   },
...
   "StoreOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name",
     "city",
     "is_active",
     "created_at"
    ],
    "title": "StoreOut"
   },
...
   "UserCreate": {
    "properties": {
     "email": {
      "type": "string",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 6,
      "title": "Password"
     },
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "role": {
      "$ref": "#/components/schemas/Role",
      "default": "OPERATOR"
     },
     "store_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Store Id"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "UserCreate"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "store_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Store Id"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     },
     "store": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StoreOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "store_id",
     "is_active",
     "created_at"
    ],
    "title": "UserOut"
   },
   "UserUpdate": {
    "properties": {
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "password": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128,
        "minLength": 6
       },
       {
        "type": "null"
       }
      ],
      "title": "Password"
     },
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Role"
       },
       {
        "type": "null"
       }
      ]
     },
     "store_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Store Id"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
 "servers": [
  {
   "url": "http://localhost:8000"
  }
 ]
}